server/requirements.txt
aiohappyeyeballs==2.3.5 ; python_version >= "3.12" and python_version < "4.0"
aiohttp==3.10.2 ; python_version >= "3.12" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.4.0 ; python_version >= "3.12" and python_version < "4.0"
attrs==24.2.0 ; python_version >= "3.12" and python_version < "4.0"
blinker==1.8.2 ; python_version >= "3.12" and python_version < "4.0"
build==1.2.1 ; python_version >= "3.12" and python_version < "4.0"
cachecontrol[filecache]==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.7.4 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.0 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
charset-normalizer==3.3.2 ; python_version >= "3.12" and python_version < "4.0"
cleo==2.1.0 ; python_version >= "3.12" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and (platform_system == "Windows" or os_name == "nt")
crashtest==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
cryptography==43.0.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
dataclasses-json==0.6.7 ; python_version >= "3.12" and python_version < "4.0"
distlib==0.3.8 ; python_version >= "3.12" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
dulwich==0.21.7 ; python_version >= "3.12" and python_version < "4.0"
fastjsonschema==2.20.0 ; python_version >= "3.12" and python_version < "4.0"
filelock==3.15.4 ; python_version >= "3.12" and python_version < "4.0"
flask-cors==4.0.1 ; python_version >= "3.12" and python_version < "4.0"
flask==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
frozenlist==1.4.1 ; python_version >= "3.12" and python_version < "4.0"
greenlet==3.0.3 ; python_version < "3.13" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version >= "3.12"
gunicorn==22.0.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.5 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.0 ; python_version >= "3.12" and python_version < "4.0"
idna==3.7 ; python_version >= "3.12" and python_version < "4.0"
installer==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.12" and python_version < "4.0"
jaraco-classes==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
jiter==0.5.0 ; python_version >= "3.12" and python_version < "4.0"
jsonpatch==1.33 ; python_version >= "3.12" and python_version < "4.0"
jsonpointer==3.0.0 ; python_version >= "3.12" and python_version < "4.0"
keyring==24.3.1 ; python_version >= "3.12" and python_version < "4.0"
langchain-community==0.2.11 ; python_version >= "3.12" and python_version < "4.0"
langchain-core==0.2.29 ; python_version >= "3.12" and python_version < "4.0"
langchain-openai==0.1.20 ; python_version >= "3.12" and python_version < "4.0"
langchain-text-splitters==0.2.2 ; python_version >= "3.12" and python_version < "4.0"
langchain==0.2.12 ; python_version >= "3.12" and python_version < "4.0"
langsmith==0.1.98 ; python_version >= "3.12" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.12" and python_version < "4.0"
marshmallow==3.21.3 ; python_version >= "3.12" and python_version < "4.0"
more-itertools==10.4.0 ; python_version >= "3.12" and python_version < "4.0"
msgpack==1.0.8 ; python_version >= "3.12" and python_version < "4.0"
multidict==6.0.5 ; python_version >= "3.12" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.12" and python_version < "4.0"
openai==1.40.3 ; python_version >= "3.12" and python_version < "4.0"
orjson==3.10.7 ; python_version >= "3.12" and python_version < "4.0"
packaging==24.1 ; python_version >= "3.12" and python_version < "4.0"
pexpect==4.9.0 ; python_version >= "3.12" and python_version < "4.0"
pkginfo==1.11.1 ; python_version >= "3.12" and python_version < "4.0"
platformdirs==4.2.2 ; python_version >= "3.12" and python_version < "4.0"
poetry-core==1.9.0 ; python_version >= "3.12" and python_version < "4.0"
poetry-plugin-export==1.8.0 ; python_version >= "3.12" and python_version < "4.0"
poetry==1.8.3 ; python_version >= "3.12" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and (sys_platform == "darwin" or sys_platform == "linux") and (sys_platform == "darwin" or platform_python_implementation != "PyPy")
pydantic-core==2.20.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.8.2 ; python_version >= "3.12" and python_version < "4.0"
pyproject-hooks==1.1.0 ; python_version >= "3.12" and python_version < "4.0"
pywin32-ctypes==0.2.2 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "4.0"
rapidfuzz==3.9.6 ; python_version >= "3.12" and python_version < "4.0"
regex==2024.7.24 ; python_version >= "3.12" and python_version < "4.0"
requests-toolbelt==1.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "linux"
shellingham==1.5.4 ; python_version >= "3.12" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
sqlalchemy==2.0.32 ; python_version >= "3.12" and python_version < "4.0"
tenacity==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
tiktoken==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
tomlkit==0.13.0 ; python_version >= "3.12" and python_version < "4.0"
tqdm==4.66.5 ; python_version >= "3.12" and python_version < "4.0"
trove-classifiers==2024.7.2 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
typing-inspect==0.9.0 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.2 ; python_version >= "3.12" and python_version < "4.0"
virtualenv==20.26.3 ; python_version >= "3.12" and python_version < "4.0"
werkzeug==3.0.3 ; python_version >= "3.12" and python_version < "4.0"
xattr==1.1.0 ; python_version >= "3.12" and python_version < "4.0" and sys_platform == "darwin"
yarl==1.9.4 ; python_version >= "3.12" and python_version < "4.0"
